加入Kame坐标系。可以使用gl.hpp中的kamePosition将其转换为OpenGL坐标。 gl.hpp中的inverseRaster函数可以将图片尺寸转换为OpenGL尺寸。 THKamechinjiru.cpp下面有一段示范程序。
程序本体：core/thkame.hpp
bbfield/bbfield.hpp是弹幕系统的头文件。现在正在考虑是否改变算法。

+++ bbfield/bbfield.hpp
@@ -14,8 +14,8 @@ float bulletRadiusSq[10] = { 1.f,1.f,1.f,1.f,1.f,1.f,1.f,1.f,1.f,1.f };
 struct Bullet
 {
 	unsigned char type;
-	int x, y; // Coordinates
-	int vx, vy; // Velocity
+	float x, y; // Coordinates
+	float vx, vy; // Velocity
 };
 
 class BBField
